cleaned code abit , added some functionality and style to product page
