+++ src/functions/products/index.js
@@ -1,0 +1,4 @@
+
+import {fetchData} from './fetch';
+
+export {fetchData};
